+++ benches/complex.rs
@@ -96,28 +96,32 @@ fn build_deep_hierarchy(taffy: &mut taffy::node::Taffy) -> taffy::node::Node {
 }
 
 fn taffy_benchmarks(c: &mut Criterion) {
-    c.bench_function("deep hierarchy - build", |b| {
+    // Increase sample size, because the examples are very small
+    let mut group = c.benchmark_group("deep hierarchy");
+    group.sample_size(200);
+
+    group.bench_function("build", |b| {
         b.iter(|| {
             let mut taffy = taffy::node::Taffy::new();
             build_deep_hierarchy(&mut taffy);
         })
     });
 
-    c.bench_function("deep hierarchy - single", |b| {
+    group.bench_function("single", |b| {
         b.iter(|| {
             let mut taffy = taffy::node::Taffy::new();
             let root = build_deep_hierarchy(&mut taffy);
-            taffy.compute_layout(root, taffy::geometry::Size::undefined()).unwrap()
+            taffy.compute_layout(root, taffy::geometry::Size::NONE).unwrap()
         })
     });
 
-    c.bench_function("deep hierarchy - relayout", |b| {
+    group.bench_function("relayout", |b| {
         let mut taffy = taffy::node::Taffy::new();
         let root = build_deep_hierarchy(&mut taffy);
 
         b.iter(|| {
             taffy.mark_dirty(root).unwrap();
-            taffy.compute_layout(root, taffy::geometry::Size::undefined()).unwrap()
+            taffy.compute_layout(root, taffy::geometry::Size::NONE).unwrap()
         })
     });
 }