+++ src/style.rs
@@ -66,18 +66,22 @@ impl Default for AlignSelf {
 #[derive(Copy, Clone, PartialEq, Eq, Debug)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub enum AlignContent {
-    /// Items are packed toward the start of the cross axis
+    /// Items are packed toward the start of the axis
     FlexStart,
-    /// Items are packed toward the end of the cross axis
+    /// Items are packed toward the end of the axis
     FlexEnd,
-    /// Items are packed along the center of the cross axis
+    /// Items are centered around the middle of the axis
     Center,
-    /// Distribute items evenly, but stretch them to fill the container
+    /// Items are stretched to fill the container
     Stretch,
-    /// Distribute items evenly, such that the first and last item are aligned with the edges
+    /// The first and last items are aligned flush with the edges of the container (no gap)
+    /// The gap between items is distributed evenly.
     SpaceBetween,
-    /// Distribute items evenly,
-    /// such that the space between items is the same as the space between the first and last item and the edges
+    /// The gap between the first and last items is exactly THE SAME as the gap between items.
+    /// The gaps are distributed evenly
+    SpaceEvenly,
+    /// The gap between the first and last items is exactly HALF the gap between items.
+    /// The gaps are distributed evenly in proportion to these ratios.
     SpaceAround,
 }
 
@@ -177,14 +181,15 @@ pub enum JustifyContent {
     FlexEnd,
     /// Items are packed along the center of the main axis
     Center,
-    /// Distribute items evenly, such that the first and last item are aligned with the edges
+    /// The first and last items are aligned flush with the edges of the container (no gap)
+    /// The gaps between items are distributed evenly.
     SpaceBetween,
-    /// Distribute items evenly,
-    /// such that the space between items is twice same as the space between the first and last item and the edges
-    SpaceAround,
-    /// Distribute items evenly,
-    /// such that the space between items is the same as the space between the first and last item and the edges
+    /// The gap between the first and last items is exactly THE SAME as the gap between items.
+    /// The gaps are distributed evenly
     SpaceEvenly,
+    /// The gap between the first and last items is exactly HALF the gap between items.
+    /// The gaps are distributed evenly in proportion to these ratios.
+    SpaceAround,
 }
 
 impl Default for JustifyContent {
@@ -285,26 +290,26 @@ impl Default for Rect<Dimension> {
 impl Rect<Dimension> {
     /// Generates a [`Rect<Dimension>`] using [`Dimension::Points`] values for `start` and `top`
     #[must_use]
-    pub fn top_from_points(start: f32, top: f32) -> Rect<Dimension> {
-        Rect { start: Dimension::Points(start), top: Dimension::Points(top), ..Default::default() }
+    pub const fn top_from_points(start: f32, top: f32) -> Rect<Dimension> {
+        Rect { start: Dimension::Points(start), top: Dimension::Points(top), ..Rect::AUTO }
     }
 
     /// Generates a [`Rect<Dimension>`] using [`Dimension::Points`] values for `end` and `bottom`
     #[must_use]
-    pub fn bot_from_points(end: f32, bottom: f32) -> Rect<Dimension> {
-        Rect { end: Dimension::Points(end), bottom: Dimension::Points(bottom), ..Default::default() }
+    pub const fn bot_from_points(end: f32, bottom: f32) -> Rect<Dimension> {
+        Rect { end: Dimension::Points(end), bottom: Dimension::Points(bottom), ..Rect::AUTO }
     }
 
     /// Generates a [`Rect<Dimension>`] using [`Dimension::Percent`] values for `start` and `top`
     #[must_use]
-    pub fn top_from_percent(start: f32, top: f32) -> Rect<Dimension> {
-        Rect { start: Dimension::Percent(start), top: Dimension::Percent(top), ..Default::default() }
+    pub const fn top_from_percent(start: f32, top: f32) -> Rect<Dimension> {
+        Rect { start: Dimension::Percent(start), top: Dimension::Percent(top), ..Rect::AUTO }
     }
 
     /// Generates a [`Rect<Dimension>`] using [`Dimension::Percent`] values for `end` and `bottom`
     #[must_use]
-    pub fn bot_from_percent(end: f32, bottom: f32) -> Rect<Dimension> {
-        Rect { end: Dimension::Percent(end), bottom: Dimension::Percent(bottom), ..Default::default() }
+    pub const fn bot_from_percent(end: f32, bottom: f32) -> Rect<Dimension> {
+        Rect { end: Dimension::Percent(end), bottom: Dimension::Percent(bottom), ..Rect::AUTO }
     }
 
     /// Generates a [`Rect<Dimension>`] using [`Dimension::Undefined`] for all values
@@ -321,7 +326,7 @@ impl Rect<Dimension> {
 
     /// Create a new Rect with [`Dimension::Points`]
     #[must_use]
-    pub fn from_points(start: f32, end: f32, top: f32, bottom: f32) -> Self {
+    pub const fn from_points(start: f32, end: f32, top: f32, bottom: f32) -> Self {
         Rect {
             start: Dimension::Points(start),
             end: Dimension::Points(end),
@@ -332,7 +337,7 @@ impl Rect<Dimension> {
 
     /// Create a new Rect with [`Dimension::Percent`]
     #[must_use]
-    pub fn from_percent(start: f32, end: f32, top: f32, bottom: f32) -> Self {
+    pub const fn from_percent(start: f32, end: f32, top: f32, bottom: f32) -> Self {
         Rect {
             start: Dimension::Percent(start),
             end: Dimension::Percent(end),
@@ -392,11 +397,11 @@ pub struct FlexboxLayout {
     pub border: Rect<Dimension>,
     /// The relative rate at which this item grows when it is expanding to fill space
     ///
-    /// 0.0 is the default value, and this value must be positive.
+    /// 0.0 is the default value, and this value must not be negative.
     pub flex_grow: f32,
     /// The relative rate at which this item shrinks when it is contracting to fit into space
     ///
-    /// 1.0 is the default value, and this value must be positive.
+    /// 1.0 is the default value, and this value must not be negative.
     pub flex_shrink: f32,
     /// Sets the initial main axis size of the item
     pub flex_basis: Dimension,
@@ -413,29 +418,34 @@ pub struct FlexboxLayout {
     pub aspect_ratio: Option<f32>,
 }
 
+impl FlexboxLayout {
+    /// The [`Default`] layout, in a form that can be used in const functions
+    pub const DEFAULT: FlexboxLayout = FlexboxLayout {
+        display: Display::Flex,
+        position_type: PositionType::Relative,
+        flex_direction: FlexDirection::Row,
+        flex_wrap: FlexWrap::NoWrap,
+        align_items: AlignItems::Stretch,
+        align_self: AlignSelf::Auto,
+        align_content: AlignContent::Stretch,
+        justify_content: JustifyContent::FlexStart,
+        position: Rect::UNDEFINED,
+        margin: Rect::UNDEFINED,
+        padding: Rect::UNDEFINED,
+        border: Rect::UNDEFINED,
+        flex_grow: 0.0,
+        flex_shrink: 1.0,
+        flex_basis: Dimension::Auto,
+        size: Size::AUTO,
+        min_size: Size::AUTO,
+        max_size: Size::AUTO,
+        aspect_ratio: None,
+    };
+}
+
 impl Default for FlexboxLayout {
     fn default() -> Self {
-        Self {
-            display: Default::default(),
-            position_type: Default::default(),
-            flex_direction: Default::default(),
-            flex_wrap: Default::default(),
-            align_items: Default::default(),
-            align_self: Default::default(),
-            align_content: Default::default(),
-            justify_content: Default::default(),
-            position: Default::default(),
-            margin: Default::default(),
-            padding: Default::default(),
-            border: Default::default(),
-            flex_grow: 0.0,
-            flex_shrink: 1.0,
-            flex_basis: Dimension::Auto,
-            size: Default::default(),
-            min_size: Default::default(),
-            max_size: Default::default(),
-            aspect_ratio: Default::default(),
-        }
+        FlexboxLayout::DEFAULT
     }
 }
 
@@ -543,6 +553,36 @@ impl FlexboxLayout {
 #[allow(clippy::bool_assert_comparison)]
 #[cfg(test)]
 mod tests {
+    use super::FlexboxLayout;
+
+    #[test]
+    fn defaults_match() {
+        let old_defaults = FlexboxLayout {
+            display: Default::default(),
+            position_type: Default::default(),
+            flex_direction: Default::default(),
+            flex_wrap: Default::default(),
+            align_items: Default::default(),
+            align_self: Default::default(),
+            align_content: Default::default(),
+            justify_content: Default::default(),
+            position: Default::default(),
+            margin: Default::default(),
+            padding: Default::default(),
+            border: Default::default(),
+            flex_grow: 0.0,
+            flex_shrink: 1.0,
+            flex_basis: super::Dimension::Auto,
+            size: Default::default(),
+            min_size: Default::default(),
+            max_size: Default::default(),
+            aspect_ratio: Default::default(),
+        };
+
+        assert_eq!(FlexboxLayout::DEFAULT, FlexboxLayout::default());
+        assert_eq!(FlexboxLayout::DEFAULT, old_defaults);
+    }
+
     mod test_flex_direction {
         use crate::style::*;