+++ RELEASES.md
@@ -6,6 +6,10 @@
 
 - Added `taffy::error::InvalidChild` Error type
 - `taffy::node::Taffy.new_leaf()` which allows the creation of new leaf-nodes without having to supply a measure function
+- Builder methods are now `const` where possible
+  - Several convenience constants have been defined: notably `FlexboxLayout::DEFAULT`
+- New `Layout` convenience constructor: `with_order(order: u32)`
+- Added support for `AlignContent::SpaceEvenly`
 
 ### 0.2.0 Changed
 
@@ -24,6 +28,7 @@
 - `taffy::error::InvalidChild` is now `taffy::error::TaffyError`
 - `taffy::error::InvalidNode` has been removed and is now just a branch on the `TaffyError` enum
 - `taffy::forest::Forest` has been merged into `taffy::node::Taffy` for a performance boost up to 90%
+- `Size::undefined()` has been removed, use `Size::NONE` instead.
 
 ### 0.2.0 Fixed